Dot thread management
If you have a thread dependent upon a single variable = true, it's easy to control the thread running by putting it in a function to run after the = true and toss a break into the ELSE which would be your end to the thread.

This way, the thread isn't always running unless they are creating a scene. Less threads = happier RedM.